ESP-IDF: build against v6.1 (CI was pinned to v5.2.7)
- CI installs ESP-IDF v6.1 instead of v5.2.7.
- Every ESP-IDF example/test app now REQUIRES the per-peripheral driver
  components it actually uses (esp_driver_i2c/spi/uart/gpio/rmt, plus
  esp_timer where a driver includes esp_timer.h). Since v6.0 the
  catch-all `driver` component no longer provides those headers.
- Config structs (i2c_master_bus_config_t, i2c_device_config_t,
  spi_bus_config_t, spi_device_interface_config_t, uart_config_t) are
  zero-initialized and assigned field by field instead of
  designated-initialized, so new struct fields in ESP-IDF releases no
  longer trip -Werror=missing-field-initializers. Drops the v5/v6 #if
  gates in the AHT21 and ENS160 tests.
- C++26 (v6.1's default) deprecates `--` and chained assignment on
  volatile: fixed in the ADE7953/AD7705/AD7706 busy-wait loops and the
  DS3231 ESP-IDF demo.
- Moved the misplaced cpp/tests/espidf/power/ade7953_test_espidf to
  cpp/tests/power/, where test_espidf.sh looks for it.
- Fixed CPP_DIR in 5 test crates and 1 example (wrong depth) and added
  the missing CPP_DIR to the APA102 apps.
- Small pre-existing fixes found along the way: ADXL362 narrowing in
  brace init, MCP4725 minimal example calling Full-only read(), BME280
  complete example using `alt` uninitialized.
- TESTING.md/AGENTS.md templates and notes updated; minimum is now
  ESP-IDF 5.3 (first release with the esp_driver_* components).

Verified locally with ESP-IDF v6.1: all 5 CI crates build; 196/225
ESP-IDF apps build. The remaining 29 failures predate v6 and are left
for the deferred all-apps CI compile work.

+++ cpp/src/chips/accelerometer/ADXL362.cpp
@@ -69,19 +69,19 @@ void ADXL362Minimal::init() {
 }
 
 void ADXL362Minimal::_write_reg(uint8_t reg, uint8_t value) {
-    uint8_t buf[3] = { CMD_WRITE_REG, reg & 0x3F, value };
+    uint8_t buf[3] = { CMD_WRITE_REG, static_cast<uint8_t>(reg & 0x3F), value };
     _connection.write(buf, 3);
 }
 
 uint8_t ADXL362Minimal::_read_reg(uint8_t reg) {
-    uint8_t cmd[2] = { CMD_READ_REG, reg & 0x3F };
+    uint8_t cmd[2] = { CMD_READ_REG, static_cast<uint8_t>(reg & 0x3F) };
     uint8_t val = 0;
     _connection.write_read(cmd, 2, &val, 1);
     return val;
 }
 
 void ADXL362Minimal::_read_burst(uint8_t reg, uint8_t* buf, uint8_t len) {
-    uint8_t cmd[2] = { CMD_READ_REG, reg & 0x3F };
+    uint8_t cmd[2] = { CMD_READ_REG, static_cast<uint8_t>(reg & 0x3F) };
     _connection.write_read(cmd, 2, buf, len);
 }
